+++ CHANGELOG.rst
@@ -5,6 +5,7 @@ Release NEXT
 ------------
 - Introduce ram, vcpu, storage quotas to service project link.
 - Drop SynchronizationStates, SynchronizationStateFilter, SynchronizableMixin.
+- Add Default Price List Items action "Reinit configurations".
 
 Release 0.134.0
 ---------------